+++ tests/setup/main.tf
@@ -1,14 +1,3 @@
-terraform {
-  required_providers {
-    random = {
-      source  = "hashicorp/random"
-      version = "3.6.3"
-    }
-  }
-}
-
-
-
 resource "random_pet" "cluster_name" {
   length = 2
 }